+++ Gemfile
@@ -1,2 +1,2 @@
-source "https://rubygems.org"
-gem "rubocop", ">= 1.0", "< 2.0"
+source 'https://rubygems.org'
+gem 'rubocop', '>= 1.0', '< 2.0'
