--- classes/Student.rb
@@ -1,12 +1,0 @@
-require_relative './Person.rb'
-
-class Student < Person
-  def initialize(age, classroom, name = 'Unknown', parent_permission = true)
-    super(age, name, parent_permission)
-    @classroom = classroom
-  end
-
-  def play_hooky
-    "¯\(ツ)/¯"
-  end
-end
