+++ classes/teacher.rb
@@ -1,0 +1,12 @@
+require_relative './person'
+
+class Teacher < Person
+  def initialize(age, specialization)
+    super(age, name, parent_permission: true)
+    @specialization = specialization
+  end
+
+  def can_use_services?
+    true
+  end
+end
